[xim, x11] correctly shutdown xim, destroy x11 with error handling
libX11 code shows that, it will send XimMessage if window
exists. (IS_SERVER_CONNECTED), urxvt (and maybe other xim implemation)
will use XDestroyIC after fcitx die, which is totally wrong, since
no one can reply that message, and it's synchronized, thus it will
cause application freeze. But there is some local memory need
to be released with XDestroyIC, so we need to make sure xim server
appears to be disconnected before IMCloseIM.

From libX11 implementation, the only way to avoid XIM api to send
message is to destroy the window first, and call IMCloseIM, which
looks ridiculous, because ximWindow is passed to IMOpenIM as a valid
parameter, destroy it before IMCloseIM doesn't follow the instinct.
But it seems that it's the only way to avoid freeze on application side.

Anyway, if fcitx is killed by signal 9, upper solution will not work
for sure (and no solution will work, for sure). Who cares? Since xim
still sucks.