[fcitx] fix crash while reload
addon pointer got changed in process event, so the loop in
RunInstance access the addon pointer is invalid.

This would only happen if someone has a lot of addons which
got enabled and disabled once, thus reloading addon might
change the size of array and realloc the memory of the array.
Since enable/disable addon and reinstall the system is quite rare,
hopefully there aren't many people hit this bug.