fcitx-configtool.desktop: Remove MimeType entry
We are encountering the following lintian warning in Debian:

W: desktop-mime-but-no-exec-code
 usr/share/applications/fcitx-configtool.desktop

It seems that fcitx-configtool does not accept filename/filelist
as parameter (in fact it is only a wrapper of concrete config tools).
In this case, we should not add any MimeType key inside desktop
file.

See https://lintian.debian.org/tags/desktop-mime-but-no-exec-code.html
for more information.